Split into files, add initial tests
